+++ src/admin-user-roles/admin-user-roles.routes.js
@@ -43,6 +43,12 @@
                 },
                 user: function(userRoleAssignmentFactory, $stateParams, roles, programs, supervisoryNodes, warehouses) {
                     return userRoleAssignmentFactory.getUser($stateParams.id, roles, programs, supervisoryNodes, warehouses);
+                },
+                roleRightsMap: function(roles) {
+                    return roles.reduce(function(roleMap, role) {
+                        roleMap[role.id] = role.rights;
+                        return roleMap;
+                    }, {});
                 }
             },
             views: {