+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Bug fixes:
   * if a user has no a home facility and some home facility roles, an error icon will be displayed next to those role assignments.
   * a user has ability to add new home facility roles but if the user has no home facility, those assignments will be ignored.
 * [OLMIS-3675](https://openlmis.atlassian.net/browse/OLMIS-3675): Fixed pagination problems for nested views.
+* [OLMIS-4830](https://openlmis.atlassian.net/browse/OLMIS-4830): Added popover with the list of rights when hover over role on the User Roles and User Profile screens.
 
 5.3.0 / 2018-04-24
 ==================